Add variable for timeout of triggered monitor
This is useful when certain metrics are only reported for limited time
and them disappearing meaning that the triggered state can be safely
resolved.

For instance, when you track the runtime of a background job and alert
if it has been running for a long time, the metrics no longer being
received probably means that the job has finished. In this you also want
the alert to be resolved.